add trait and implementation
Add `ToGron` trait and implement it for `serde::Value` and
`rustc_serialize::Json`. Add a convenient `json_to_gron` function for
emulating the old implementation.

Test section checks both `ro_gron` implementations.
